Refactors comment modal into its own seperate component
